# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: URL Shortener API
  description: |
    A comprehensive API for URL shortening with analytics, user management, subscriptions, and payment processing.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or HTTP-only cookies.
    - **Access Token**: Short-lived token for API authentication
    - **Refresh Token**: Long-lived token for obtaining new access tokens

    ## Rate Limiting
    API requests may be rate limited based on your subscription plan.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5080/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /urls:
    get:
      tags:
        - URLs
      summary: Get user's URLs
      description: Get all shortened URLs for the authenticated user with pagination and filtering.
      operationId: getMyUrls
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          description: Search term for URL or alias
          schema:
            type: string
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: string
            enum: ["true", "false"]
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [createdAt, clicks, originalUrl]
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: URLs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - URLs
      summary: Create short URL
      description: Create a new shortened URL.
      operationId: createUrl
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUrlRequest"
      responses:
        "201":
          description: URL created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Custom alias already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /urls/summary:
    get:
      tags:
        - URLs
      summary: Get URL summary
      description: Get summary statistics for user's URLs.
      operationId: getMyUrlSummary
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /urls/{id}:
    get:
      tags:
        - URLs
      summary: Get URL by ID
      description: Get details of a specific shortened URL.
      operationId: getUrlById
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UrlId"
      responses:
        "200":
          description: URL retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags:
        - URLs
      summary: Update URL
      description: Update a shortened URL's settings.
      operationId: updateUrl
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UrlId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUrlRequest"
      responses:
        "200":
          description: URL updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - URLs
      summary: Delete URL
      description: Delete a shortened URL.
      operationId: deleteUrl
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UrlId"
      responses:
        "200":
          description: URL deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /urls/admin/all:
    get:
      tags:
        - URLs
      summary: Get all URLs (Admin)
      description: Get all URLs in the system. Admin only.
      operationId: getAllUrls
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          description: Search term
          schema:
            type: string
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: URLs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /urls/admin/{id}:
    get:
      tags:
        - URLs
      summary: Get URL by ID (Admin)
      description: Get any URL by ID. Admin only.
      operationId: adminGetUrlById
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UrlId"
      responses:
        "200":
          description: URL retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - URLs
      summary: Delete URL (Admin)
      description: Delete any URL by ID. Admin only.
      operationId: adminDeleteUrl
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UrlId"
      responses:
        "200":
          description: URL deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== ANALYTICS/CLICKS ROUTES ====================
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users (Admin)
      description: Get all users with pagination and filtering. Admin only.
      operationId: getAllUsers
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
            maxLength: 100
        - name: role
          in: query
          description: Filter by user role
          schema:
            type: string
            enum: [admin, user]
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Users
      summary: Create user (Admin)
      description: Create a new user. Admin only.
      operationId: createUser
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID (Admin)
      description: Get a user by their ID. Admin only.
      operationId: getUserById
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Users
      summary: Update user (Admin)
      description: Update a user's information. Admin only.
      operationId: updateUser
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Users
      summary: Delete user (Admin)
      description: Delete a user. Admin only.
      operationId: deleteUser
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== REDIRECT ROUTE ====================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HTTP-only cookie containing JWT access token

  # ==================== PARAMETERS ====================
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    UrlId:
      name: id
      in: path
      required: true
      description: URL ID
      schema:
        type: string
    UserId:
      name: id
      in: path
      required: true
      description: User ID
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Authentication required
            statusCode: 401
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Access denied. Admin privileges required.
            statusCode: 403
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Resource not found
            statusCode: 404
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ==================== SCHEMAS ====================
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        statusCode:
          type: integer
          example: 400
      required:
        - success
        - message
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Invalid email format
      required:
        - success
        - message
        - errors
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false

    # Auth Schemas
...
    User:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum: [admin, user]
          example: user
        avatar:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        isActive:
          type: boolean
          example: true
        isEmailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/User"
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[a-zA-Z\\s]+$"
          example: John
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[a-zA-Z\\s]+$"
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          minLength: 6
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)"
          example: Password123
        role:
          type: string
          enum: [admin, user]
          example: user
        isActive:
          type: boolean
          example: true
        isEmailVerified:
          type: boolean
          example: false
      required:
        - firstName
        - lastName
        - email
        - password
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[a-zA-Z\\s]+$"
          example: John
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[a-zA-Z\\s]+$"
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum: [admin, user]
          example: user
        isActive:
          type: boolean
          example: true
        isEmailVerified:
          type: boolean
          example: true
        avatar:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
...
    Url:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        originalUrl:
          type: string
          format: uri
          example: https://example.com/very/long/url/path
        shortId:
          type: string
          example: abc123
        shortUrl:
          type: string
          format: uri
          example: https://short.url/abc123
        user:
          type: string
          example: 507f1f77bcf86cd799439011
        clickCount:
          type: integer
          example: 150
        isActive:
          type: boolean
          example: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-12-31T23:59:59Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    CreateUrlRequest:
      type: object
      properties:
        originalUrl:
          type: string
          format: uri
          example: https://example.com/very/long/url/path
        customAlias:
          type: string
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9_-]+$"
          example: my-custom-link
          description: Optional custom alias (3-50 characters, alphanumeric, hyphens, underscores)
        expiresAt:
          type: string
          format: date-time
          example: "2024-12-31T23:59:59Z"
          description: Optional expiration date
      required:
        - originalUrl
    UpdateUrlRequest:
      type: object
      properties:
        originalUrl:
          type: string
          format: uri
          example: https://example.com/updated/url
        isActive:
          type: boolean
          example: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-12-31T23:59:59Z"
    UrlResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Url"
    UrlListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Url"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    UrlSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            totalUrls:
              type: integer
              example: 50
            activeUrls:
              type: integer
              example: 45
            totalClicks:
              type: integer
              example: 5000
            todayClicks:
              type: integer
              example: 120

    # Click/Analytics Schemas
...